requirements.txt
# Core dependencies
flask==3.0.0
pymongo==4.6.0
python-dotenv==1.0.0

# AI/ML Libraries
google-generativeai==0.3.2
langgraph==0.0.26
langchain-core==0.1.25

# Document Processing
pypdf==3.17.4
python-docx==1.1.0

# Data Processing
datasets==2.16.1
sentence-transformers==2.2.2
faiss-cpu==1.9.0.post1

# Additional utilities
requests==2.31.0
pydantic==2.9.2
